API Reference

Pre-Built Architectures

NeuralOperators.NOMAD — Type

NOMAD(approximator, decoder, concatenate)

Constructs a NOMAD from approximator and decoder architectures. Make sure the output from approximator combined with the coordinate dimension has compatible size for input to decoder

Arguments

approximator: Lux network to be used as approximator net.

decoder: Lux network to be used as decoder net.

Keyword Arguments

concatenate: function that defines the concatenation of output from approximator and the coordinate dimension, defaults to concatenation along first dimension after vectorizing the tensors

References

[1] Jacob H. Seidman and Georgios Kissas and Paris Perdikaris and George J. Pappas, "NOMAD: Nonlinear Manifold Decoders for Operator Learning", doi: https://arxiv.org/abs/2206.03551

Example

julia> approximator_net = Chain(Dense(8 => 32), Dense(32 => 32), Dense(32 => 16));

julia> decoder_net = Chain(Dense(18 => 16), Dense(16 => 16), Dense(16 => 8));

julia> nomad = NOMAD(approximator_net, decoder_net);

julia> ps, st = Lux.setup(Xoshiro(), nomad);

julia> u = rand(Float32, 8, 5);

julia> y = rand(Float32, 2, 5);

julia> size(first(nomad((u, y), ps, st)))
(8, 5)

NeuralOperators.DeepONet — Type

DeepONet(branch, trunk, additional)

Constructs a DeepONet from a branch and trunk architectures. Make sure that both the nets output should have the same first dimension.

Arguments

branch: Lux network to be used as branch net.

trunk: Lux network to be used as trunk net.

Keyword Arguments

additional: Lux network to pass the output of DeepONet, to include additional operations for embeddings, defaults to nothing

References

[1] Lu Lu, Pengzhan Jin, George Em Karniadakis, "DeepONet: Learning nonlinear operators for identifying differential equations based on the universal approximation theorem of operators", doi: https://arxiv.org/abs/1910.03193

Example

julia> branch_net = Chain(Dense(64 => 32), Dense(32 => 32), Dense(32 => 16));

julia> trunk_net = Chain(Dense(1 => 8), Dense(8 => 8), Dense(8 => 16));

julia> deeponet = DeepONet(branch_net, trunk_net);

julia> ps, st = Lux.setup(Xoshiro(), deeponet);

julia> u = rand(Float32, 64, 5);

julia> y = rand(Float32, 1, 10, 5);

julia> size(first(deeponet((u, y), ps, st)))
(10, 5)

NeuralOperators.FourierNeuralOperator — Type

FourierNeuralOperator(
    σ=gelu; chs::Dims{C}=(2, 64, 64, 64, 64, 64, 128, 1), modes::Dims{M}=(16,),
    permuted::Val{perm}=False, kwargs...) where {C, M, perm}

Fourier neural operator is a operator learning model that uses Fourier kernel to perform spectral convolutions. It is a promising way for surrogate methods, and can be regarded as a physics operator.

The model is comprised of a Dense layer to lift (d + 1)-dimensional vector field to n-dimensional vector field, and an integral kernel operator which consists of four Fourier kernels, and two Dense layers to project data back to the scalar field of interest space.

Arguments

σ: Activation function for all layers in the model.

Keyword Arguments

chs: A Tuple or Vector of the 8 channel size.

modes: The modes to be preserved. A tuple of length d, where d is the dimension of data.

permuted: Whether the dim is permuted. If permuted = Val(false), the layer accepts data in the order of (ch, x_1, ... , x_d , batch). Otherwise the order is (x_1, ... , x_d, ch, batch).

Example

julia> fno = FourierNeuralOperator(gelu; chs=(2, 64, 64, 128, 1), modes=(16,));

julia> ps, st = Lux.setup(Xoshiro(), fno);

julia> u = rand(Float32, 2, 1024, 5);

julia> size(first(fno(u, ps, st)))
(1, 1024, 5)

Building blocks

NeuralOperators.OperatorConv — Type

OperatorConv(ch::Pair{<:Integer, <:Integer}, modes::Dims,
    ::Type{<:AbstractTransform}; init_weight=glorot_uniform,
    permuted=Val(false))

Arguments

ch: A Pair of input and output channel size ch_in => ch_out, e.g. 64 => 64.

modes: The modes to be preserved. A tuple of length d, where d is the dimension of data.

::Type{TR}: The transform to operate the transformation.

Keyword Arguments

init_weight: Initial function to initialize parameters.

permuted: Whether the dim is permuted. If permuted = Val(false), the layer accepts data in the order of (ch, x_1, ... , x_d, batch). Otherwise the order is (x_1, ... , x_d, ch, batch).

Example

julia> OperatorConv(2 => 5, (16,), FourierTransform{ComplexF32});

julia> OperatorConv(2 => 5, (16,), FourierTransform{ComplexF32}; permuted=Val(true));

NeuralOperators.SpectralConv — Function

SpectralConv(args...; kwargs...)

Construct a OperatorConv with FourierTransform{ComplexF32} as the transform. See OperatorConv for the individual arguments.

Example

julia> SpectralConv(2 => 5, (16,));

julia> SpectralConv(2 => 5, (16,); permuted=Val(true));

NeuralOperators.OperatorKernel — Type

OperatorKernel(ch::Pair{<:Integer, <:Integer}, modes::Dims, transform::Type{TR},
    act::A=identity; permuted=Val(false), kwargs...) where {TR <: AbstractTransform, A}

Arguments

ch: A Pair of input and output channel size ch_in => ch_out, e.g. 64 => 64.

modes: The modes to be preserved. A tuple of length d, where d is the dimension of data.

::Type{TR}: The transform to operate the transformation.

Keyword Arguments

σ: Activation function.

permuted: Whether the dim is permuted. If permuted = Val(true), the layer accepts data in the order of (ch, x_1, ... , x_d , batch). Otherwise the order is (x_1, ... , x_d, ch, batch).

All the keyword arguments are passed to the OperatorConv constructor.

Example

julia> OperatorKernel(2 => 5, (16,), FourierTransform{ComplexF64});

julia> OperatorKernel(2 => 5, (16,), FourierTransform{ComplexF64}; permuted=Val(true));

NeuralOperators.SpectralKernel — Function

SpectralKernel(args...; kwargs...)

Construct a OperatorKernel with FourierTransform{ComplexF32} as the transform. See OperatorKernel for the individual arguments.

Example

julia> SpectralKernel(2 => 5, (16,));

julia> SpectralKernel(2 => 5, (16,); permuted=Val(true));

Transform API

NeuralOperators.AbstractTransform — Type

AbstractTransform

Interface

Base.ndims(<:AbstractTransform): N dims of modes

transform(<:AbstractTransform, x::AbstractArray): Apply the transform to x

truncate_modes(<:AbstractTransform, x_transformed::AbstractArray): Truncate modes that contribute to the noise

inverse(<:AbstractTransform, x_transformed::AbstractArray): Apply the inverse transform to x_transformed